feat(routes): /annotation-campaign/$campaignID/phase/$phaseType/edit-annotators
