+++ frontend/src/routes/_authenticated/annotation-campaign/$campaignID._detailLayout.module.scss
