+++ frontend/src/routes/_authenticated/annotation-campaign/$campaignID._detailLayout/phase.$phaseType.lazy.tsx
