+++ frontend/src/routes/_authenticated/annotation-campaign/$campaignID._detailLayout/phase.$phaseType.module.scss
