+++ frontend/src/routes/_authenticated/annotation-campaign/$campaignID.phase.$phaseType/edit-annotators.lazy.tsx
@@ -1,0 +1,233 @@
+import React, { Fragment, useCallback, useEffect, useMemo, useState } from 'react';
+import { createLazyFileRoute, useRouter } from '@tanstack/react-router';
+import { IonButton, IonNote, IonSkeletonText, IonSpinner } from '@ionic/react';
+import { QueryStatus } from '@reduxjs/toolkit/query';
+
+import { GraphQLErrorText, Head, Table, Tbody, Th, Thead, Tr, useToast } from '@/components/ui';
+import { FormBloc, type Item, ListSearchbar, type SearchItem } from '@/components/form';
+
+import {
+    AnnotationFileRangeInput,
+    useAllFileRanges,
+    useAllUsers,
+    useCurrentCampaign,
+    useCurrentPhase,
+    useUpdateFileRanges,
+} from '@/api';
+import { getNewItemID } from '@/service/function';
+
+import { FileRangeInputRow } from '@/features/AnnotationFileRange';
+
+import styles from './edit-annotators.module.scss';
+
+type FileRange = Omit<AnnotationFileRangeInput, 'id'> & {
+    id: string;
+    started?: boolean;
+}
+
+const EditAnnotators: React.FC = () => {
+    const { phaseType } = Route.useParams();
+    const {
+        campaign,
+        isFetching: isFetchingCampaign,
+        error: errorLoadingCampaign,
+    } = useCurrentCampaign();
+    const { phase } = useCurrentPhase()
+    const router = useRouter();
+    const toast = useToast();
+    const {
+        users, groups,
+        isFetching: isFetchingUsers,
+        error: errorLoadingUsers,
+    } = useAllUsers()
+    const {
+        allFileRanges,
+        isFetching: isFetchingFileRanges,
+        error: errorLoadingFileRanges,
+    } = useAllFileRanges()
+    const {
+        updateFileRanges,
+        isLoading: isSubmitting,
+        error: errorSubmitting,
+        formErrors,
+        status: submissionStatus,
+    } = useUpdateFileRanges()
+    const [ force, setForce ] = useState<boolean>()
+
+    // File ranges
+    const [ fileRanges, setFileRanges ] = useState<FileRange[]>([]);
+    const availableUsers: SearchItem[] = useMemo(() => {
+        const items: SearchItem[] = [];
+        if (users) {
+            items.push(...users.filter(u => {
+                if (!campaign?.spectrogramsCount) return true;
+                const count = fileRanges
+                    .filter(f => f.annotatorId === u!.id)
+                    .reduce((count, range) => {
+                        const last_index = range.lastFileIndex ?? campaign.spectrogramsCount ?? 0;
+                        const first_index = range.firstFileIndex ?? 0;
+                        return count + (last_index - first_index)
+                    }, 0) + 1
+                return count < campaign.spectrogramsCount
+            }).map(u => ({
+                value: `user-${ u!.id }`,
+                label: u!.displayName || u!.username,
+                searchable: [ ...u!.displayName!.split(' '), u!.username ],
+            } as SearchItem)));
+        }
+        if (groups) {
+            items.push(...groups.map(g => ({
+                value: `group-${ g!.id }`,
+                label: g!.name,
+                searchable: [ g!.name ],
+            } as SearchItem)))
+        }
+        return items;
+    }, [ users, campaign, fileRanges, groups ]);
+    useEffect(() => {
+        if (allFileRanges) setFileRanges(allFileRanges.map(r => ({
+            id: r!.id,
+            annotatorId: r!.annotator.id,
+            firstFileIndex: r!.firstFileIndex,
+            lastFileIndex: r!.lastFileIndex,
+            started: !!r!.completedAnnotationTasks?.totalCount,
+        })));
+    }, [ allFileRanges ]);
+    const addFileRange = useCallback((item: Item) => {
+        if (!groups || !campaign?.spectrogramsCount) return;
+        const [ type, id ] = (item.value as string).split('-');
+        const newUsers: any[] = []
+        switch (type!) {
+            case 'user':
+                newUsers.push(users.find(a => a!.id === id)!);
+                break;
+            case 'group':
+                newUsers.push(...groups.find(g => g!.id === id)!.users!.filter(u => availableUsers.find(a => a.value.split('-')[0] === 'user' && a.value.split('-')[1] === u?.id)));
+                break
+        }
+        setFileRanges(prev => {
+            for (const newUser of newUsers) {
+                prev = [ ...prev, {
+                    id: getNewItemID(prev)?.toString(),
+                    annotatorId: newUser!.id,
+                    firstFileIndex: 1,
+                    lastFileIndex: campaign.spectrogramsCount,
+                } ]
+            }
+            return prev
+        })
+    }, [ users, groups, availableUsers, setFileRanges, campaign ])
+    const updateFileRange = useCallback((fileRange: FileRange) => {
+        setFileRanges(prev => prev.map(f => {
+            if (f.id !== fileRange.id) return f;
+            return { ...f, ...fileRange }
+        }))
+    }, [])
+    const removeFileRange = useCallback((fileRange: FileRange) => {
+        setFileRanges(prev => prev.filter(f => f.id !== fileRange.id))
+    }, [])
+
+    // Navigation
+    const back = useCallback(() => router.history.back(), [ router ])
+
+    // Submit
+    const submit = useCallback(() => {
+        updateFileRanges({ fileRanges, force })
+    }, [ fileRanges, updateFileRanges, force ])
+    useEffect(() => {
+        if (errorSubmitting) toast.raiseError({ error: errorSubmitting })
+    }, [ errorSubmitting ]);
+    useEffect(() => {
+        if (submissionStatus === QueryStatus.fulfilled) back()
+    }, [ submissionStatus ]);
+
+    return useMemo(() =>
+            <Fragment>
+
+                <Head title="Manage annotators"
+                      subtitle={ campaign ? `${ campaign.name } - ${ phaseType }` :
+                          <IonSkeletonText animated style={ { width: 128 } }/> }/>
+
+                <FormBloc className={ styles.annotators }>
+
+                    <ListSearchbar placeholder="Search annotator..."
+                                   disabled={ isFetchingCampaign || isFetchingUsers || isFetchingFileRanges }
+                                   values={ availableUsers }
+                                   onValueSelected={ addFileRange }/>
+
+                    {/* Loading */ }
+                    { (isFetchingCampaign || isFetchingUsers || isFetchingFileRanges) && <IonSpinner/> }
+                    { errorLoadingCampaign &&
+                        <GraphQLErrorText error={ errorLoadingCampaign }/> }
+                    { errorLoadingUsers &&
+                        <GraphQLErrorText error={ errorLoadingUsers }/> }
+                    { errorLoadingFileRanges &&
+                        <GraphQLErrorText error={ errorLoadingFileRanges }/> }
+
+                    { !(isFetchingCampaign || isFetchingUsers || isFetchingFileRanges) &&
+                        <Table>
+                            <Thead>
+                                <Tr>
+                                    <Th scope="col">Annotator</Th>
+                                    <Th scope="col" colSpan={ 2 }>
+                                        File range
+                                        <br/>
+                                        <small>(between 1 and { campaign?.spectrogramsCount })</small>
+                                        <br/>
+                                        <small className="disabled"><i>Start and end limits are included</i></small>
+                                    </Th>
+                                    <Th scope="col"/>
+                                </Tr>
+                            </Thead>
+                            <Tbody>
+                                { fileRanges.map((range, k) => {
+                                        const user = users.find(u => u?.id == range.annotatorId)
+                                        if (!user) return <Fragment/>
+                                        return <FileRangeInputRow key={ k }
+                                                                  range={ range }
+                                                                  errors={ formErrors[k] ?? undefined }
+                                                                  annotator={ user }
+                                                                  onUpdate={ change => {
+                                                                      updateFileRange({
+                                                                          ...range,
+                                                                          ...change,
+                                                                      })
+                                                                  } }
+                                                                  setForced={ () => setForce(true) }
+                                                                  onDelete={ removeFileRange }/>
+                                    },
+                                ) }
+                            </Tbody>
+
+                            { fileRanges.length === 0 && <IonNote color="medium">No annotators</IonNote> }
+                        </Table>
+                    }
+
+                    { phase?.phase === 'Verification' &&
+                        <IonNote>To fully verify your annotations, you should have a verification user that is not an
+                            annotator
+                            or
+                            at
+                            least two verification users</IonNote> }
+
+                    <div className={ styles.buttons }>
+                        <IonButton color="medium" fill="outline" onClick={ back }>
+                            Back to campaign
+                        </IonButton>
+                        { isSubmitting && <IonSpinner/> }
+                        <IonButton disabled={ isSubmitting } onClick={ submit }>
+                            Update annotators
+                        </IonButton>
+                    </div>
+
+                </FormBloc>
+            </Fragment>,
+        [ campaign, phaseType, addFileRange, errorLoadingCampaign, isFetchingUsers, formErrors, errorLoadingUsers,
+            submit, isSubmitting, isFetchingCampaign, fileRanges, phase, users, isFetchingFileRanges, availableUsers,
+            back, errorLoadingFileRanges ],
+    )
+}
+
+export const Route = createLazyFileRoute('/_authenticated/annotation-campaign/$campaignID/phase/$phaseType/edit-annotators')({
+    component: EditAnnotators,
+})
