+++ frontend/src/routes/_authenticated/annotation-campaign/$campaignID.phase.$phaseType/edit-annotators.module.scss
